TIP-863: move Group to Enrichment
