+++ src/Pim/Bundle/CatalogBundle/Resources/config/entities.yml
@@ -10,7 +10,7 @@ parameters:
     pim_catalog.entity.association_type_translation.class: Pim\Bundle\CatalogBundle\Entity\AssociationTypeTranslation
     pim_catalog.entity.category.class:                     Akeneo\Pim\Enrichment\Component\Category\Model\Category
     pim_catalog.entity.category_translation.class:         Akeneo\Pim\Enrichment\Component\Category\Model\CategoryTranslation
-    pim_catalog.entity.group.class:                        Pim\Bundle\CatalogBundle\Entity\Group
+    pim_catalog.entity.group.class:                        Akeneo\Pim\Enrichment\Component\Product\Model\Group
     pim_catalog.entity.group_translation.class:            Pim\Bundle\CatalogBundle\Entity\GroupTranslation
     pim_catalog.entity.group_type.class:                   Pim\Bundle\CatalogBundle\Entity\GroupType
     pim_catalog.entity.group_type_translation.class:       Pim\Bundle\CatalogBundle\Entity\GroupTypeTranslation
