+++ tests/back/Acceptance/Catalog/InMemoryGroupRepository.php
@@ -7,7 +7,7 @@ namespace Akeneo\Test\Acceptance\Catalog;
 use Akeneo\Tool\Component\StorageUtils\Saver\SaverInterface;
 use Akeneo\Test\Acceptance\Common\NotImplementedException;
 use Doctrine\Common\Collections\ArrayCollection;
-use Pim\Bundle\CatalogBundle\Entity\Group;
+use Akeneo\Pim\Enrichment\Component\Product\Model\Group;
 use Akeneo\Pim\Enrichment\Component\Product\Model\GroupInterface;
 use Pim\Component\Catalog\Repository\GroupRepositoryInterface;
